Split modal into subcomponents
